requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pyyaml==6.0.2
pydantic==2.9.2
sse-starlette==2.1.3
openai>=1.30.0
